% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcht.R
\name{addcht}
\alias{addcht}
\title{Add a new factor with a change between the values of two factors in a tidy table}
\usage{
addcht(df, VarA, VarB, Var, Value)
}
\arguments{
\item{...}{numeric}
}
\value{
a dataframe
}
\description{
This function is used to calculate changes between values of categories (dates, ages, countries, etc).
}
\details{
The data may come as TIDY, meaning that all values are in 1 column and other columns show information related to the value, meaning categories.

All inputs must be in quotes ("").

\strong{Inputs:}
\enumerate{
 \item  df   =  is the database containing a column with Factors.
 \item  Var1 =  ARE the "VariableS" valueS that will be used to calculate the differences vs Var0.
 \item  Var0 =  IS the Variable value  anchor for all changes calculation.
}
Var1, Var0 and Var must be factors with leverls or characters NOT NUMBERS
}
\examples{

}
