% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrank.R
\name{addrank}
\alias{addrank}
\title{Add a "MyRank" column with the rank of a variable}
\usage{
addrank(df, Var)
}
\arguments{
\item{...}{numeric}
}
\value{
a dataframe
}
\description{
Add a "MyRank" column with the rank of a variable
}
\details{
This function needs the following imputs:
\enumerate{
 \item \strong{df}  = dataframe.
 \item \strong{Var} = variables for which the rank will be added.  e.g. the df has 2 categories in a column that we will need to add individual ranks, we include the variables quoted of in that column that will be ranked.
}
}
\examples{

}
