% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtasbr.R
\name{addtasbr}
\alias{addtasbr}
\title{This is a get total and subtotal by rank addtasbr}
\usage{
addtasbr(df, Rank, RankT, ...)
}
\arguments{
\item{...}{numeric}
}
\value{
a dataframe
}
\description{
This is a get total and subtotal by rank addtasbr
}
\examples{

}
