% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addw.R
\name{addw}
\alias{addw}
\title{This function calculates create a weighted variable}
\usage{
addw(df, Var, Weight, FilterVarCero = FALSE, Group = FALSE)
}
\arguments{
\item{...}{numeric}
}
\value{
a dataframe
}
\description{
This function calculates create a weighted variable
}
\examples{

}
