% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettotalenq.R
\name{gettotalenq}
\alias{gettotalenq}
\title{This is a get total when columns are in quotes}
\usage{
gettotalenq(df, ...)
}
\arguments{
\item{...}{numeric}
}
\value{
a dataframe
}
\description{
This is a get total when columns are in quotes
}
\examples{

}
