% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettotalsym.R
\name{gettotalsym}
\alias{gettotalsym}
\title{This is a get total when columns are in symbors}
\usage{
gettotalsym(df, ...)
}
\arguments{
\item{...}{numeric}
}
\value{
a dataframe
}
\description{
This is a get total when columns are in symbors
}
\examples{

}
